+++ service/internal/pipelines/pipelines.go
@@ -355,7 +355,7 @@ func buildExporter(
 		BuildInfo:         buildInfo,
 	}
 	set.TelemetrySettings.Logger = exporterLogger(settings.Logger, id, pipelineID.Type())
-	components.LogStabilityLevel(set.TelemetrySettings.Logger, factory.StabilityLevel(pipelineID.Type()))
+	components.LogStabilityLevel(set.TelemetrySettings.Logger, getExporterStabilityLevel(factory, pipelineID.Type()))
 
 	exp, err := createExporter(ctx, set, cfg, id, pipelineID, factory)
 	if err != nil {
@@ -413,6 +413,18 @@ func exporterLogger(logger *zap.Logger, id config.ComponentID, dt config.DataTyp
 		zap.String(components.ZapNameKey, id.String()))
 }
 
+func getExporterStabilityLevel(factory component.ExporterFactory, dt config.DataType) component.StabilityLevel {
+	switch dt {
+	case config.TracesDataType:
+		return factory.TracesExporterStability()
+	case config.MetricsDataType:
+		return factory.MetricsExporterStability()
+	case config.LogsDataType:
+		return factory.LogsExporterStability()
+	}
+	return component.StabilityLevelUndefined
+}
+
 func buildProcessor(ctx context.Context,
 	settings component.TelemetrySettings,
 	buildInfo component.BuildInfo,
@@ -437,7 +449,7 @@ func buildProcessor(ctx context.Context,
 		BuildInfo:         buildInfo,
 	}
 	set.TelemetrySettings.Logger = processorLogger(settings.Logger, id, pipelineID)
-	components.LogStabilityLevel(set.TelemetrySettings.Logger, factory.StabilityLevel(pipelineID.Type()))
+	components.LogStabilityLevel(set.TelemetrySettings.Logger, getProcessorStabilityLevel(factory, pipelineID.Type()))
 
 	proc, err := createProcessor(ctx, set, procCfg, id, pipelineID, next, factory)
 	if err != nil {
@@ -467,6 +479,18 @@ func processorLogger(logger *zap.Logger, procID config.ComponentID, pipelineID c
 		zap.String(components.ZapKindPipeline, pipelineID.String()))
 }
 
+func getProcessorStabilityLevel(factory component.ProcessorFactory, dt config.DataType) component.StabilityLevel {
+	switch dt {
+	case config.TracesDataType:
+		return factory.TracesProcessorStability()
+	case config.MetricsDataType:
+		return factory.MetricsProcessorStability()
+	case config.LogsDataType:
+		return factory.LogsProcessorStability()
+	}
+	return component.StabilityLevelUndefined
+}
+
 func buildReceiver(ctx context.Context,
 	settings component.TelemetrySettings,
 	buildInfo component.BuildInfo,
@@ -491,7 +515,7 @@ func buildReceiver(ctx context.Context,
 		BuildInfo:         buildInfo,
 	}
 	set.TelemetrySettings.Logger = receiverLogger(settings.Logger, id, pipelineID.Type())
-	components.LogStabilityLevel(set.TelemetrySettings.Logger, factory.StabilityLevel(pipelineID.Type()))
+	components.LogStabilityLevel(set.TelemetrySettings.Logger, getReceiverStabilityLevel(factory, pipelineID.Type()))
 
 	recv, err := createReceiver(ctx, set, cfg, id, pipelineID, nexts, factory)
 	if err != nil {
@@ -530,6 +554,18 @@ func receiverLogger(logger *zap.Logger, id config.ComponentID, dt config.DataTyp
 		zap.String(components.ZapKindKey, components.ZapKindReceiver),
 		zap.String(components.ZapNameKey, id.String()),
 		zap.String(components.ZapKindPipeline, string(dt)))
+}
+
+func getReceiverStabilityLevel(factory component.ReceiverFactory, dt config.DataType) component.StabilityLevel {
+	switch dt {
+	case config.TracesDataType:
+		return factory.TracesReceiverStability()
+	case config.MetricsDataType:
+		return factory.MetricsReceiverStability()
+	case config.LogsDataType:
+		return factory.LogsReceiverStability()
+	}
+	return component.StabilityLevelUndefined
 }
 
 func (bps *Pipelines) getPipelinesSummaryTableData() zpages.SummaryPipelinesTableData {